% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportAmibroker.R
\name{ImportAmibroker}
\alias{ImportAmibroker}
\title{Title}
\usage{
ImportAmibroker(data, tickerlist)
}
\arguments{
\item{tickerlist}{}
}
\description{
Title
}
\examples{
dataraw <- read.table("data-raw/FULLDATA20132018.txt", sep = ",", header=T)
tickerlist <- data(tickerlist2)
ImportAmibroker(dataraw, tickerlist) # Auto assignm variables.
}
