% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportAmibroker.R
\name{ImportAmibrokerDF}
\alias{ImportAmibrokerDF}
\title{Title}
\usage{
ImportAmibrokerDF(data, tickerlist)
}
\arguments{
\item{tickerlist}{}
}
\description{
Title
}
\examples{
#' dataraw <- read.table("data-raw/FULLDATA20132018.txt", sep = ",", header=T)
tickerlist <- data(tickerlist2)
final_data <- ImportAmibroker(dataraw, tickerlist)
}
