% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAmibroker-package.R
\docType{package}
\name{RAmibroker-package}
\alias{RAmibroker}
\alias{RAmibroker-package}
\title{RAmibroker: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
